// estore/src/app/Shared/models/products.ts
export default class Product{
  Code: string;
  ProductName: string;
  ProductCat: string;
  ProductDis: string;
  ProductPrice: number;
  ProductQuantity: number;
  AddedBy: string;
  DateAdded: string;
  ImagePath: string;
  ImagePath2: string;
  Size: Array<any>;
  Color: Array<any>;
  Manufacturer: string;
}